Transition to Silk.NET for OpenAl, SDL2 and ShaderC
Transition Headless to Silk.NET

Resource Managment

Transition Audio

Naming changes

prevent glsl gc collection

Silk.net Shaderc

Sdl.Free use here is illegal

+++ src/Ryujinx.Headless.SDL2/Program.cs
@@ -527,7 +527,7 @@ namespace Ryujinx.Headless.SDL2
 
                 return new VulkanRenderer(
                     api,
-                    (instance, vk) => new SurfaceKHR((ulong)(vulkanWindow.CreateWindowSurface(instance.Handle))),
+                    (instance, vk) => new SurfaceKHR((vulkanWindow.CreateWindowSurface(instance.Handle))),
                     vulkanWindow.GetRequiredInstanceExtensions,
                     preferredGpuId);
             }
